Move cookie signing code into separate file to make it easier to share
