+++ lib/appSession.js
@@ -1,7 +1,5 @@
 const { strict: assert, AssertionError } = require('assert');
 const {
-  JWK,
-  JWKS,
   JWE,
   errors: { JOSEError },
 } = require('jose');
@@ -9,7 +7,7 @@ const { promisify } = require('util');
 const cookie = require('cookie');
 const onHeaders = require('on-headers');
 const COOKIES = require('./cookies');
-const { encryption: deriveKey } = require('./hkdf');
+const { getKeyStore } = require('./crypto');
 const debug = require('./debug')('appSession');
 
 const epoch = () => (Date.now() / 1000) | 0;
@@ -47,13 +45,8 @@ function replaceSession(req, session, config) {
 }
 
 module.exports = (config) => {
-  let current;
-
   const alg = 'dir';
   const enc = 'A256GCM';
-  const secrets = Array.isArray(config.secret)
-    ? config.secret
-    : [config.secret];
   const sessionName = config.session.name;
   const cookieConfig = config.session.cookie;
   const {
@@ -74,16 +67,7 @@ module.exports = (config) => {
   );
   const cookieChunkSize = MAX_COOKIE_SIZE - emptyCookie.length;
 
-  let keystore = new JWKS.KeyStore();
-
-  secrets.forEach((secretString, i) => {
-    const key = JWK.asKey(deriveKey(secretString));
-    if (i === 0) {
-      current = key;
-    }
-    keystore.add(key);
-  });
-
+  let [current, keystore] = getKeyStore(config.secret, true);
   if (keystore.size === 1) {
     keystore = current;
   }
