+++ lib/transientHandler.js
@@ -1,66 +1,14 @@
 const { generators } = require('openid-client');
-const { JWKS, JWS, JWK } = require('jose');
-const { signing: deriveKey } = require('./hkdf');
-
-const header = { alg: 'HS256', b64: false, crit: ['b64'] };
-const getPayload = (cookie, value) => Buffer.from(`${cookie}=${value}`);
-const flattenedJWSFromCookie = (cookie, value, signature) => ({
-  protected: Buffer.from(JSON.stringify(header))
-    .toString('base64')
-    .replace(/=/g, '')
-    .replace(/\+/g, '-')
-    .replace(/\//g, '_'),
-  payload: getPayload(cookie, value),
-  signature,
-});
-const generateSignature = (cookie, value, key) => {
-  const payload = getPayload(cookie, value);
-  return JWS.sign.flattened(payload, key, header).signature;
-};
-const verifySignature = (cookie, value, signature, keystore) => {
-  try {
-    return !!JWS.verify(
-      flattenedJWSFromCookie(cookie, value, signature),
-      keystore,
-      { algorithm: 'HS256', crit: ['b64'] }
-    );
-  } catch (err) {
-    return false;
-  }
-};
-const getCookieValue = (cookie, value, keystore) => {
-  if (!value) {
-    return undefined;
-  }
-  let signature;
-  [value, signature] = value.split('.');
-  if (verifySignature(cookie, value, signature, keystore)) {
-    return value;
-  }
-
-  return undefined;
-};
-
-const generateCookieValue = (cookie, value, key) => {
-  const signature = generateSignature(cookie, value, key);
-  return `${value}.${signature}`;
-};
-
+const {
+  signCookie: generateCookieValue,
+  verifyCookie: getCookieValue,
+  getKeyStore,
+} = require('./crypto');
 const COOKIES = require('./cookies');
 
 class TransientCookieHandler {
   constructor({ secret, session, legacySameSiteCookie }) {
-    let current;
-
-    const secrets = Array.isArray(secret) ? secret : [secret];
-    let keystore = new JWKS.KeyStore();
-    secrets.forEach((secretString, i) => {
-      const key = JWK.asKey(deriveKey(secretString));
-      if (i === 0) {
-        current = key;
-      }
-      keystore.add(key);
-    });
+    let [current, keystore] = getKeyStore(secret);
 
     if (keystore.size === 1) {
       keystore = current;
